chore: Move submodule source out of module data into its own module
