+++ hipcheck/src/data/git/query/mod.rs
@@ -17,8 +17,8 @@ use crate::data::git::RawCommit;
 use crate::data::git::SignerKeyView;
 use crate::data::git::SignerNameView;
 use crate::data::git::SignerView;
-use crate::data::source::SourceQuery;
 use crate::error::Result;
+use crate::source::source::SourceQuery;
 use crate::version::VersionQuery;
 use chrono::prelude::*;
 use std::rc::Rc;
