+++ hipcheck/src/data/query/code_quality.rs
@@ -6,8 +6,8 @@ use std::rc::Rc;
 
 use crate::data::code_quality::get_eslint_report;
 use crate::data::code_quality::CodeQualityReport;
-use crate::data::source::SourceQuery;
 use crate::error::Result;
+use crate::source::source::SourceQuery;
 use crate::version::VersionQuery;
 
 /// Queries about code quality
