+++ hipcheck/src/session/session.rs
@@ -19,12 +19,6 @@ use crate::context::Context as _;
 use crate::data::git::get_git_version;
 use crate::data::git::GitProviderStorage;
 use crate::data::npm::get_npm_version;
-use crate::data::source::Source;
-use crate::data::source::SourceChangeRequest;
-use crate::data::source::SourceKind;
-use crate::data::source::SourceQuery;
-use crate::data::source::SourceQueryStorage;
-use crate::data::source::SourceRepo;
 use crate::data::CodeQualityProviderStorage;
 use crate::data::DependenciesProviderStorage;
 use crate::data::FuzzProviderStorage;
@@ -45,6 +39,12 @@ use crate::session::pm::detect_and_extract;
 use crate::session::spdx::extract_download_url;
 use crate::shell::Phase;
 use crate::shell::Shell;
+use crate::source::source::Source;
+use crate::source::source::SourceChangeRequest;
+use crate::source::source::SourceKind;
+use crate::source::source::SourceQuery;
+use crate::source::source::SourceQueryStorage;
+use crate::source::source::SourceRepo;
 use crate::version::get_version;
 use crate::version::VersionQuery;
 use crate::version::VersionQueryStorage;
