+++ hipcheck/src/source.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod query;
+#[allow(clippy::module_inception)]
+pub mod source;
